--- mensa_notfications.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-from mensa import send_notifications
-
-if __name__ == '__main__':
-    send_notifications()
